minor doc fixes
